claude_agent_course/module_08_capstone_project/projects/devmate/pyproject.toml
[project]
name = "devmate"
version = "0.1.0"
description = "DevMate: Your Autonomous AI Developer Agent"
authors = [
    {name = "Claude Agent Course", email = "course@example.com"},
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

dependencies = [
    "anthropic>=0.30.0",
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.0.0",
    "pathspec>=0.11.0"  # For parsing .gitignore
]

[project.scripts]
devmate = "src.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
